Address startup issues, attempt to identify misconfiguration issues with proper errors
