+++ letsencrypt/client/errors.py
@@ -5,6 +5,10 @@ class LetsEncryptClientError(Exception):
     """Generic Let's Encrypt client error."""
 
 
+class LetsEncryptReverterError(LetsEncryptClientError):
+    """Let's Encrypt Reverter error."""
+
+
 class LetsEncryptAuthHandlerError(LetsEncryptClientError):
     """Let's Encrypt Auth Handler error."""
 
@@ -15,6 +19,10 @@ class LetsEncryptClientAuthError(LetsEncryptAuthHandlerError):
 
 class LetsEncryptConfiguratorError(LetsEncryptClientError):
     """Let's Encrypt Configurator error."""
+
+
+class LetsEncryptMisconfigurationError(LetsEncryptClientError):
+    """Let's Encrypt Misconfiguration Error."""
 
 
 class LetsEncryptDvsniError(LetsEncryptConfiguratorError):
